Change: Put all compilation option structs to opts module
