+++ src/compile/build.rs
@@ -1,13 +1,9 @@
 use bytes::Bytes;
 use failure::Error;
-use inkwell::targets::{
-    CodeModel, FileType, InitializationConfig, RelocMode, Target, TargetMachine,
-};
-use inkwell::OptimizationLevel;
-use structopt::clap::{_clap_count_exprs, arg_enum};
-use structopt::StructOpt;
+use inkwell::targets::{CodeModel, FileType, InitializationConfig, Target, TargetMachine};
 
 use super::llvm;
+use super::opts::{BuildOpt, OutputType};
 use crate::error::{CLIError, LLVMError};
 use crate::parser;
 use crate::transform::TransformManager;
@@ -15,102 +11,6 @@ use crate::translator::translate_ast;
 
 use std::fs::File;
 use std::io::{Read, Write};
-use std::path::PathBuf;
-
-arg_enum! {
-    enum OutputType {
-        Object,
-        Assembly,
-        IR,
-        EIR,
-        AST,
-    }
-}
-
-arg_enum! {
-    enum OptimizationLevelOpt {
-        None,
-        Less,
-        Default,
-        Aggressive,
-    }
-}
-
-impl Into<OptimizationLevel> for OptimizationLevelOpt {
-    fn into(self) -> OptimizationLevel {
-        match self {
-            OptimizationLevelOpt::None => OptimizationLevel::None,
-            OptimizationLevelOpt::Less => OptimizationLevel::Less,
-            OptimizationLevelOpt::Default => OptimizationLevel::Default,
-            OptimizationLevelOpt::Aggressive => OptimizationLevel::Aggressive,
-        }
-    }
-}
-
-arg_enum! {
-    enum RelocModeOpt {
-        Default,
-        Static,
-        PIC,
-        DynamicNoPIC,
-    }
-}
-
-impl Into<RelocMode> for RelocModeOpt {
-    fn into(self) -> RelocMode {
-        match self {
-            RelocModeOpt::Default => RelocMode::Default,
-            RelocModeOpt::Static => RelocMode::Static,
-            RelocModeOpt::PIC => RelocMode::PIC,
-            RelocModeOpt::DynamicNoPIC => RelocMode::DynamicNoPic,
-        }
-    }
-}
-
-#[derive(StructOpt)]
-pub struct CodegenOpt {
-    #[structopt(long = "triple")]
-    target_triple: Option<String>,
-
-    #[structopt(long = "cpu")]
-    target_cpu: Option<String>,
-
-    #[structopt(long = "cpu-features")]
-    target_cpu_features: Option<String>,
-
-    #[structopt(short = "O", long = "optimize", default_value = "default")]
-    #[structopt(raw(
-        possible_values = "&OptimizationLevelOpt::variants()",
-        case_insensitive = "true"
-    ))]
-    optimization_level: OptimizationLevelOpt,
-
-    #[structopt(long = "reloc", default_value = "default")]
-    #[structopt(raw(
-        possible_values = "&RelocModeOpt::variants()",
-        case_insensitive = "true"
-    ))]
-    reloc_mode: RelocModeOpt,
-
-    #[structopt(long = "emit-func-name", default_value = "main")]
-    emit_func_name: String,
-}
-
-#[derive(StructOpt)]
-pub struct BuildOpt {
-    #[structopt(name = "FILE", parse(from_os_str))]
-    input: PathBuf,
-
-    #[structopt(short = "o", long = "output", parse(from_os_str))]
-    output: PathBuf,
-
-    #[structopt(short = "t", long = "output-type", default_value = "object")]
-    #[structopt(raw(possible_values = "&OutputType::variants()", case_insensitive = "true"))]
-    output_type: OutputType,
-
-    #[structopt(flatten)]
-    codegen_opt: CodegenOpt,
-}
 
 pub fn build(opt: BuildOpt) -> Result<(), Error> {
     let BuildOpt {
