+++ src/compile/mod.rs
@@ -1,3 +1,4 @@
 pub mod build;
 pub mod jit;
 pub mod llvm;
+pub mod opts;
